requirements.txt
psycopg2-binary==2.9.1
gunicorn==20.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decouple==0.0.7
dnspython==2.1.0
dominate==2.5.2
email-validator==1.1.3
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.4.1
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
is-disposable-email==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-decouple==3.4
python-dotenv==0.19.0
requests==2.24.0
SQLAlchemy==1.3.19
urllib3==1.25.10
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
wtforms-validators==1.0.0
gunicorn==20.1.0
appdirs==1.4.4
apturl==0.5.2
bcrypt==3.1.7
beautifulsoup4==4.9.3
blinker==1.4
Brlapi==0.7.0
bs4==0.0.1
cachelib==0.3.0
cachetools==4.0.0
cchardet==2.1.7
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
colorgram.py==1.2.0
command-not-found==0.3
commonmark==0.9.1
cryptography==2.8
cssselect==1.1.0
cupshelpers==1.0
dbus-python==1.2.16
decorator==5.0.9
defer==1.0.6
Deprecated==1.2.12
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.1.0
duplicity==0.8.12.0
email-validator==1.1.3
emoji==1.4.2
emojis==0.6.0
entrypoints==0.3
evdev==1.3.0
fasteners==0.14.1
Flask==2.0.1
Flask-WTF==0.15.1
freegames==2.3.2
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.0.3
heroes==1.0.2
howdoi==2.0.17
html2text==2020.1.16
html5lib==1.0.1
htmlfetcher==0.0.6
htmlparsing==0.1.5
httplib2==0.14.0
idna==2.8
iterfzf==0.5.0.20.0
itsdangerous==2.0.1
Jinja2==3.0.1
keep==2.10.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
monotonic==1.5
netifaces==0.10.4
ntlm-auth==1.5.0
numpy==1.21.1
oauth2client==4.1.2
oauthlib==3.1.0
olefile==0.46
pandas==1.3.0
paramiko==2.6.0
parse==1.19.0
pbr==5.6.0
pexpect==4.6.0
Pillow==8.3.1
protobuf==3.6.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
PyDrive==1.3.1
PyGithub==1.55
Pygments==2.9.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyOpenSSL==19.0.0
pyperclip==1.8.2
pyquery==1.4.3
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-decouple==3.4
python-docx==0.8.11
python-dotenv==0.19.0
python-magic==0.4.16
PythonTurtle==0.3.2
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
qrcode==7.2
ranger-fm==1.9.3
reportlab==3.5.34
requests==2.26.0
requests-ntlm==1.1.0
requests-unixsocket==0.2.0
rich==10.6.0
rsa==4.0
sanitize-filename==1.2.0
SecretStorage==2.3.1
selenium==3.141.0
self==2020.12.3
setproctitle==1.1.10
simplejson==3.16.0
six==1.16.0
soupsieve==2.2.1
speedtest-cli==2.1.2
spotipy==2.19.0
SQLAlchemy==1.3.12
systemd-python==234
terminaltables==3.1.0
testresources==2.0.1
tkintertable==1.3.3
toapi==2.1.3
toml==0.10.2
tqdm==4.61.2
twilio==6.63.0
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unicodecsv==0.14.1
uritemplate==0.6
urllib3==1.26.6
usb-creator==0.3.7
wadllib==1.3.3
webencodings==0.5.1
Werkzeug==2.0.1
wikipedia==1.4.0
wrapt==1.12.1
WTForms==2.3.3
xkit==0.0.0


